+++ DICOM/IO/ByteConverter.cs
@@ -54,8 +54,8 @@ namespace Dicom.IO
 #else
             uint size = (uint)Marshal.SizeOf(typeof(T));
 #endif
-            uint padding = buffer.Size % size;
-            uint count = buffer.Size / size;
+            uint padding = (uint)(buffer.Size % size);
+            uint count = (uint)(buffer.Size / size);
             T[] values = new T[count];
             System.Buffer.BlockCopy(buffer.Data, 0, values, 0, (int)(buffer.Size - padding));
             return values;
